// src/shared/models/BoxData.ts
export interface BoxData {
    id: number
    time: string
    temp: number
    temp_feel: number
    desc: string
    icon: string
    selected: boolean
    rainfall_risk: number
    rainfall_amount: number
    wind_speed: number
    wind_direction: string
    wind_gust_speed: number
    wind_gust_direction: string
    sunshine_duration_in_min: number
    sunshine_uv_index: number
    humidity_in_percent: number
    pressure_in_hpa: number


}